monitor/requirements.txt
paho-mqtt==1.6.1
python-dotenv==0.20.0